% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_remove.R
\name{object_remove}
\alias{object_remove}
\title{object_remove}
\usage{
object_remove()
}
\description{
Removes objects from the global environment according to a pattern for easy use as an RStudio addin
}
\examples{
if (interactive()) {
  object_remove()
}
}
